+++ Example/Emission/index.storybooks.js
@@ -3,7 +3,7 @@ import { getStorybookUI, configure } from '@kadira/react-native-storybook';
 
 // import your stories
 configure(() => {
-  require("../../src/storiesRoot")
+  require("../../src/storiesRegistry")
 }, module);
 
 const StorybookUI = getStorybookUI({port: 9001, host: 'localhost'});