+++ src/lib/components/buttons/__stories__/buttons.story.tsx
@@ -2,9 +2,6 @@ import { action, storiesOf } from "@kadira/react-native-storybook"
 import * as React from "react"
 import { View } from "react-native"
 
-import { RootContainer } from "react-relay"
-import StubContainer from "react-storybooks-relay-container"
-
 import FlatWhite from "../flat_white"
 import InvertedButton from "../inverted_button"
 import NavigationButton from "../navigation_button"
@@ -12,7 +9,7 @@ import NavigationButton from "../navigation_button"
 const smallButton = { height: 26, width: 320, marginBottom: 20 }
 const largeButton = { height: 26, width: 320, marginBottom: 20 }
 
-storiesOf("Buttons")
+storiesOf("🎨 Buttons")
   .addDecorator((story) => (
     <View style={{ marginTop: 60, marginLeft: 20, marginRight: 20 }}>{story()}</View>
   ))