+++ src/lib/components/consignments/__tests__/index-tests.tsx
@@ -1,0 +1,9 @@
+import * as React from "react"
+import "react-native"
+import * as renderer from "react-test-renderer"
+import Consignments from "../"
+
+it("Sets up the right view heirarchy", () => {
+    const tree = renderer.create(<Consignments />).toJSON()
+    expect(tree).toMatchSnapshot()
+})