+++ src/lib/components/consignments/setup/__tests__/welcome-tests.tsx
@@ -1,0 +1,12 @@
+import * as React from "react"
+import "react-native"
+import * as renderer from "react-test-renderer"
+import Welcome from "../welcome"
+
+it("Sets up the right view heirarchy", () => {
+      const nav = {} as any
+      const route = {} as any
+
+      const tree = renderer.create(<Welcome navigator={nav} route={route}/>).toJSON()
+      expect(tree).toMatchSnapshot()
+})