+++ src/lib/components/text/__stories__/typography.story.tsx
@@ -1,0 +1,13 @@
+import { storiesOf } from "@kadira/react-native-storybook"
+import * as React from "react"
+
+import Headline from "../headline"
+import Serif from "../serif"
+
+storiesOf("🎨 Typography")
+  .add("App Headline", () => {
+    return <Headline>This is a balnk headline</Headline>
+  })
+  .add("App Serif Text", () => {
+    return <Serif>This is a blank serif</Serif>
+  })